Renames port finding functions for clarity.
The original names of `findAnOpenPort` and `findAClosedPort` were ambiguous.
While the functions were referring to port status, open could be interpreted as
available. New functions are `findAPortInUse` and `findAPortNotInUse`. Verbose,
yes.  I am open to suggestions that keep the same tense and verb usage. So, not
`findAnAvailablePort` and `findAPortInUse`, for example.